fix(oxygen): replace double with decimal for all monetary/rate values
ScoreTier.BaseRate, ScoreTier.MaxLoanPercentage, TermMultiplier.Value,
LoanApplication.Amount, ApplyForLoanRequest.LoanAmount, ILoanService
signature — all changed from double to decimal. Amortisation math
now uses DecimalPower.Pow (binary exponentiation) instead of Math.Pow
to avoid binary floating-point rounding errors.

+++ oxygen/Oxygen.Service/Interfaces/ILoanService.cs
@@ -4,5 +4,5 @@ namespace Oxygen.Service;
 
 public interface ILoanService
 {
-    Task<LoanApplicationDTO> ApplyForLoan(string userId, double loanAmount, int term);
+    Task<LoanApplicationDTO> ApplyForLoan(string userId, decimal loanAmount, int term);
 }
